Add imports

+++ src/parser/symbols/expressions/literals/IntegerLiteral.java
@@ -1,5 +1,7 @@
 package parser.symbols.expressions.literals;
 
+import dot.DotNode;
+
 public final class IntegerLiteral extends Literal {
     
     public IntegerLiteral(int value) {
@@ -21,7 +23,7 @@ public final class IntegerLiteral extends Literal {
         DotNode dotNode = new DotNode(buffer, "INT_LIT", "", "filled", "#00a2ff");
         
         dotNode.addEdge((StringBuilder buffer1) -> {
-            DotNode dotNode1 = new DotNode(buffer1, Integer.toString(value), "plaintext", "", "");
+            DotNode dotNode1 = new DotNode(buffer1, Integer.toString((int) value), "plaintext", "", "");
         }, "ident");
     }
